Refactor form stories
