--- old-storybook/src/stories/Label.stories.js
@@ -1,29 +1,0 @@
-import { createLabel } from './Form/Form.js'
-import FormDocs from './Form/Form.mdx'
-
-// More on default export: https://storybook.js.org/docs/html/writing-stories/introduction#default-export
-export default {
-  title: 'Components/Form/Label',
-  // More on argTypes: https://storybook.js.org/docs/html/api/argtypes
-  argTypes: {
-    label: { control: 'text' },
-  },
-  parameters: {
-    docs: {
-      page: FormDocs,
-    },
-  },
-}
-
-// More on component templates: https://storybook.js.org/docs/html/writing-stories/introduction#using-args
-const Template = ({ label, ...args }) => {
-  // You can either use a function to create DOM elements or use a plain html string!
-  // return `<div>${label}</div>`;
-  return createLabel({ label, ...args })
-}
-
-export const Default = Template.bind({})
-// More on args: https://storybook.js.org/docs/html/writing-stories/args
-Default.args = {
-  label: 'A Label',
-}
