--- old-storybook/src/stories/Select.stories.js
@@ -1,36 +1,0 @@
-import { createSelect } from './Form/Form.js'
-import FormDocs from './Form/Form.mdx'
-
-export default {
-  title: 'Components/Form/Select',
-  argTypes: {
-    options: {
-      control: {
-        type: 'number',
-        min: 1,
-        max: 5,
-      },
-    },
-    readonly: {
-      control: { type: 'boolean' },
-    },
-    multiple: {
-      control: { type: 'boolean' },
-    },
-  },
-  parameters: {
-    docs: {
-      page: FormDocs,
-    },
-  },
-}
-
-const Template = ({ options, ...args }) => {
-  return createSelect({ options, ...args })
-}
-
-export const Default = Template.bind({})
-Default.args = {
-  options: 1,
-  multiple: false,
-}
