+++ src/stories/Components/Form/Form.mdx
@@ -1,123 +1,98 @@
-import { Primary, ArgsTable, Canvas, Story } from '@storybook/addon-docs'
-import LinkTo from '@storybook/addon-links/react'
+import { Meta, Story, Canvas, Controls } from '@storybook/blocks'
+import * as ErrorSummaryStories from './ErrorSummary.stories'
+import * as FormGroupStories from './FormGroup.stories'
+import * as InputStories from './Input.stories'
+import * as LabelStories from './Label.stories'
+import * as SelectStories from './Select.stories'
+import * as TextareaStories from './Textarea.stories'
+
+<Meta title="Components/Form" />
 
 # Form
 
 Form classes can be used on a variety of `inputs` or `select` html elements.
 They provide consistent and composable styling that should address most applications basic needs.
 
-<!-- Either use <Stories /> to automatically render each story -->
-<!-- Or manually render them with custom descriptions -->
+## Playground
+
+<Canvas of={FormGroupStories.Default} />
+<Controls of={FormGroupStories.Default} />
 
 ## Label
 
 `.form-label` Provides basic input label styles.
 
-<Canvas withToolbar>
-  <Story id="components-form-label--default" />
-</Canvas>
+<Canvas of={LabelStories.Default} />
 
 ## Input
 
 `.form-control` Provides basic input styles. This can be used on `input` html elements. It works across a variety of types such as `text`, `number`, `password`, `email`, `phone`, `color`, `radio`, and, `checkbox`
 
-<Canvas withToolbar>
-  <Story id="components-form-input--default" />
-</Canvas>
+<Canvas of={InputStories.Default} />
 
 ## Form Group
 
 `.form-group` Composes an input and label together as a pair. It provides padding between input + label pairs in a form.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--default" />
-</Canvas>
+<Canvas of={FormGroupStories.Default} />
 
 ## Textarea
 
 `textarea.form-control` Provides basic text area styles. This can be used on `textarea` html elements.
 
-<Canvas withToolbar>
-  <Story id="components-form-textarea--default" />
-</Canvas>
+<Canvas of={TextareaStories.Default} />
 
 ## Dropdown
 
 `select.form-control` Provides basic `select` element styles.
 
-<Canvas withToolbar>
-  <Story id="components-form-select--default" />
-</Canvas>
+<Canvas of={SelectStories.Default} />
 
 ## Color
 
 `<input type='color' class='form-control'>` Provides basic color input styles. This can be used on `input` html elements with a type of `color`.
 
-<Canvas withToolbar>
-  <Story id="components-form-input--color" />
-</Canvas>
+<Canvas of={InputStories.Color} />
 
 ## Radio
 
 `<input type='radio' class='form-control'>` Provides basic input of type `radio` styles.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--radio" />
-</Canvas>
+<Canvas of={FormGroupStories.Radio} />
 
 ## Checkbox
 
 `<input type='checkbox' class='form-control'>` Provides basic input of type `radio` styles.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--checkbox" />
-</Canvas>
+<Canvas of={FormGroupStories.Checkbox} />
 
 ## Error Summary
 
 `.form-error-summary` Is an error section, usually displayed above the form with a list of errors.
 
-<Canvas withToolbar>
-  <Story id="components-form-errorsummary--default" />
-</Canvas>
+<Canvas of={ErrorSummaryStories.Default} />
 
 ## Input Error
 
 `.form-group--error` and `.form-error` are typically used on combination when there is a validation error on an input. They highlight the input in error along with the message provided.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--error" />
-</Canvas>
+<Canvas of={FormGroupStories.Error} />
 
 ## Input Hint
 
 `.form-hint` can be used inside of a `.form-group` to provide help text for an input.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--hint" />
-</Canvas>
+<Canvas of={FormGroupStories.Hint} />
 
 ## Read Only
 
 `.form-control--read-only` on any input with `.form-control` can be used to display values in a form-like way.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--readonly" />
-</Canvas>
+<Canvas of={FormGroupStories.Readonly} />
 
 ## Indeterminate Checkbox
 
 Checkboxes that use the indeterminate psuedo-class with be displayed with a `-` inside the checkmark.
 See [MDN's documentation](https://developer.mozilla.org/en-US/docs/Web/CSS/:indeterminate) for further information.
 
-<Canvas withToolbar>
-  <Story id="components-form-formgroup--indeterminate-checkbox" />
-</Canvas>
-
-## Playground
-
-Shows the default example of the currently selected component
-
-<Primary />
-
-<ArgsTable story="^" />
+<Canvas of={FormGroupStories.IndeterminateCheckbox} />
