+++ src/stories/Components/Form/Label.stories.js
@@ -1,0 +1,17 @@
+import { createLabel } from './Form.js'
+
+export default {
+  title: 'Components/Form/Label',
+  render: ({ label, ...args }) => {
+    return createLabel({ label, ...args })
+  },
+  argTypes: {
+    label: { control: 'text' },
+  },
+}
+
+export const Default = {
+  args: {
+    label: 'A Label',
+  },
+}
