+++ src/stories/Components/Form/Textarea.stories.js
@@ -1,0 +1,17 @@
+import { createTextarea } from './Form.js'
+
+export default {
+  title: 'Components/Form/Textarea',
+  render: ({ ...args }) => {
+    return createTextarea({ ...args })
+  },
+  argTypes: {
+    readonly: {
+      control: { type: 'boolean' },
+    },
+  },
+}
+
+export const Default = {
+  args: {},
+}
